updated ranking method to the advised scenarios and cleaned up the code
